1. add windows dx9-dx12 backend support 2. fixed imnodes win32 build issue 3. add all demo to directx11 and glfw

+++ addon/imguiDock/imgui_dock.cpp
@@ -1058,7 +1058,7 @@ struct DockContext
         //SetNextWindowSize(ImVec2(300, 300));
         if (Begin("Dock Debug Info")) {
             for (int i = 0; i < m_docks.size(); ++i) {
-                if (TreeNode((const void*)(long)i, "Dock %d (%p)", i, m_docks[i])) {
+                if (TreeNode((const void*)(long long)i, "Dock %d (%p)", i, m_docks[i])) {
                     Dock &dock = *m_docks[i];
                     Text("pos=(%.1f %.1f) size=(%.1f %.1f)", 
                         dock.pos.x, dock.pos.y,
